backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase==2.3.0
psycopg2-binary==2.9.9
pgvector==0.2.4
asyncpg==0.29.0

# LangChain and LangGraph
langchain==0.1.6
langchain-core==0.1.23
langchain-openai==0.0.2
langchain-community==0.0.20
langchain-google-genai==0.0.6
langchain-groq==0.1.1
langgraph==0.0.26
langsmith==0.0.87

# Embeddings
openai==1.12.0
cohere==4.47

# MLOps
mlflow==2.11.1
evidently==0.4.15
ragas==0.1.7
datasets==2.16.1

# Background jobs
prefect==2.14.11
celery==5.3.6
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
PyJWT[crypto]>=2.10.1
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
# NumPy - use version with pre-built wheels (Python 3.13 compatible)
numpy>=2.0.0
pandas>=2.0.0
tiktoken==0.5.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3

